# docs/openapi-3.0.yaml
openapi: 3.0.0
info:
  contact:
    email: support@volumeviz.io
    name: API Support
    url: https://github.com/mantonx/volumeviz/issues
  description: >-
    Docker volume monitoring API with comprehensive volume discovery, size
    calculation, and container attachment tracking. Focus on user-mounted
    volumes only.


    ## Volume-First Approach

    - Automatic discovery and filtering of user-mounted volumes

    - Excludes Docker infrastructure volumes (container filesystems, tmp volumes)

    - Real-time volume usage monitoring and size calculation

    - Container attachment tracking for each volume


    ## Features

    - Multi-method volume size calculation (du, find, stat)

    - Asynchronous scanning with progress tracking for large volumes

    - Real-time WebSocket progress updates during scan operations

    - High-performance caching with TTL-based invalidation

    - Comprehensive Prometheus metrics integration

    - Circuit breaker patterns for resilience


    ## WebSocket API

    WebSocket endpoint: `ws://localhost:8080/ws`


    ### Scan Progress Events

    Subscribe to real-time scan progress updates:

    ```json

    {

    "type": "subscribe",

    "event": "scan_progress",

    "data": {

    "volume_id": "volume_id_here",

    "scan_id": "scan_id_here"

    }

    }

    ```


    Progress updates are broadcast as:

    ```json

    {

    "type": "scan_progress_update",

    "data": {

    "scan_id": "scan_id",

...

...

...

    "phases": [

    {

...

...

...
paths:
...
  /api/v1/organizations/me:
    get:
      description: Get the organization details for the authenticated user
      responses:
        "200":
          description: Organization details with stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_organizations.OrganizationWithStats\
                  Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
      summary: Get current organization
      tags:
        - organizations
    put:
      description: Update organization details (admin only)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_api_v1_organizations.UpdateOrganizationReq\
                uest"
        description: Organization update details
        required: true
      responses:
        "200":
          description: Updated organization details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_organizations.OrganizationResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "403":
          description: Forbidden - admin required
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
      summary: Update current organization
      tags:
        - organizations
...
servers:
...
components:
  schemas:
...
    internal_api_v1_organizations.OrganizationResponse:
      properties:
        created_at:
          example: 2024-01-01T12:00:00Z
          type: string
        description:
          example: A leading technology company
          type: string
        display_name:
          example: ACME Corporation
          type: string
        id:
          example: 1
          type: integer
        is_active:
          example: true
          type: boolean
        max_storage_gb:
          example: 1000
          type: integer
        max_users:
          example: 50
          type: integer
        max_volumes:
          example: 100
          type: integer
        name:
          example: acme-corp
          type: string
        plan_type:
          example: pro
          type: string
        subdomain:
          example: acme
          type: string
        updated_at:
          example: 2024-01-01T14:30:00Z
          type: string
      type: object
...